rename changelogs to dbmigration
